+++ src/components/course/assistant-panel/PartListSection.tsx
@@ -1,10 +1,9 @@
 import { cn } from '@/lib/utils';
 import { useState } from 'react';
 import { ChevronsDown, ChevronsUp } from 'lucide-react';
-import { MOCK_PARTS } from './mockData';
 import type { Part } from './types';
-import { ASSETS } from '@/constants/assets';
-import ModelThumbnail from '@/components/common/ModelThumbnail';
+import { useCourseParts } from '@/hooks/useCourseParts';
+import PartItem from './PartItem';
 
 interface PartListSectionProps {
   selectedPartId: string | null;
@@ -12,46 +11,21 @@ interface PartListSectionProps {
   courseId?: string;
 }
 
-// Map course IDs to ASSETS keys
-const COURSE_ID_MAP: Record<string, string> = {
-  '1': 'Quadcopter_DRONE',
-  '2': 'LEAF_SPRING',
-  '3': 'MACHINE_VICE',
-  '4': 'ROBOT_ARM',
-  '5': 'ROBOT_GRIPPER',
-  '6': 'SUSPENSION',
-  '7': 'V4_ENGINE',
-};
-
 /**
  * 부품 리스트 섹션 컴포넌트
  * - 단일 책임: 부품 목록 표시 및 확장/축소 기능만 담당
+ * - 비즈니스 로직 분리: 부품 데이터 로딩 로직을 useCourseParts 훅으로 분리
  */
 export default function PartListSection({
   selectedPartId,
-  parts,
+  parts: passedParts,
   courseId,
 }: PartListSectionProps) {
   const [isExpanded, setIsExpanded] = useState(false);
-
-  // Derive parts from ASSETS if not provided
-  const displayParts: Part[] = parts ?? (() => {
-    // Map the numeric course ID to the asset key
-    const assetKey = courseId ? COURSE_ID_MAP[courseId] : undefined;
-    
-    if (!assetKey || !ASSETS[assetKey]) return MOCK_PARTS;
-    
-    return ASSETS[assetKey].modelUrls.map((url) => {
-      // Extract name from file path: "/models/drone/Arm gear.glb" -> "Arm gear"
-      const fileName = url.split('/').pop() || '';
-      const name = fileName.replace('.glb', '');
-      return {
-        id: name, // Use name as ID for now to match selection logic if possible, or url
-        name: name,
-        image: url, // We use the model URL as the 'image' source
-      };
-    });
-  })();
+  const { parts: hookParts } = useCourseParts(courseId);
+  
+  // 외부에서 주입된 parts가 있으면 사용, 없으면 hook에서 가져온 parts 사용
+  const displayParts = passedParts ?? hookParts;
 
   return (
     <section>
@@ -93,49 +67,5 @@ export default function PartListSection({
         </button>
       </div>
     </section>
-  );
-}
-
-/**
- * 개별 부품 아이템 컴포넌트
- * - 단일 책임: 개별 부품의 표시만 담당
- */
-interface PartItemProps {
-  part: Part;
-  isSelected: boolean;
-}
-
-function PartItem({ part, isSelected }: PartItemProps) {
-  const isGlb = part.image.endsWith('.glb');
-
-  return (
-    <button className="group flex w-full shrink-0 flex-col items-center gap-1 focus:outline-none">
-      <div
-        className={cn(
-          'aspect-square w-full overflow-hidden rounded-lg transition-all',
-          isSelected
-            ? 'bg-blue-50 shadow-sm ring-1 ring-blue-500'
-            : 'bg-white hover:bg-foundation-blue-1'
-        )}
-      >
-        {isGlb ? (
-          <ModelThumbnail modelUrl={part.image} />
-        ) : (
-          <img
-            src={part.image}
-            alt={part.name}
-            className="h-full w-full object-cover"
-          />
-        )}
-      </div>
-      <span
-        className={cn(
-          'w-full truncate text-center text-[10px] transition-colors',
-          isSelected ? 'font-medium text-blue-600' : 'text-gray-500 group-hover:text-active group-hover:font-medium'
-        )}
-      >
-        {part.name}
-      </span>
-    </button>
   );
 }
